backend/requirements.txt
blinker==1.9.0
certifi==2023.7.22
click==8.2.0
django-bootstrap==0.2.4
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
setuptools==68.2.2
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
wheel==0.41.3